a better way to handle incorrect /Length attribute
The old version could choke on the 
'endit = source.multiple(2)'
line because PdfTokens.multiple() accepts a valid
stream, which should not be expected the case when
the /Length parameter is broken. This problem may
happen when fdata[endstream:] starts with a closing
bracket or curly brace etc. Note in place of the 
'except:' line also 'except AssertionError:' would 
also work. But perhaps PdfTokens.broken() will
throw something other than AssertionError so it
may better to catch everything until then.